+++ changelog.md
@@ -66,6 +66,7 @@ Proof-of-concept version.
 * Naft output follows types and variables as chai
 * Added chai functionality for `add_file` and `add_key_value`
 * Added chai `getenv()` function
+* Resolving may fail if no chef is present
 
 ## Next